pyproject.toml
[tool.poetry]
name = "easyflaskapp"
version = "0.1.0"
description = "A ready to use boilerplate Flask App"
authors = ["Khieu Nguyen <nvkhieu89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pyotp = "^2.8.0"
flask = "^2.2.2"
sqlalchemy = {extras = ["asyncio"], version = "^1.4.45"}
python-dotenv = "^0.21.0"
flask-login = "^0.6.2"
mysqlclient = "^2.1.1"
gunicorn = "^20.1.0"
pyqrcode = "^1.2.1"


[tool.poetry.group.dev.dependencies]
sqlacodegen = {version = "^2.3.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
